feat(shadow-mem): MemmorySSA based on shadow-mem computed def-use
The goal is to expose the def-use chains computed by shadow-mem pass through a
an API similar to llvm::MemorySSA.

The llvm::MemorySSA provides similar information to ShadowMem, except that it
allows for a single (implicit) memory region. This patch ports the data
structures from llvm::MemorySSA into ShadowMem.

Currently, the basic data structures are ported and an algorithm to compute a
SeaMemorySSA object is implemented. The long term plan is to switch all clients
to use SeaMemorySSA and only introduce shadow.mem() annotations during building
of MemSSA, and remove them right after.